% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{post_uwebasr}
\alias{post_uwebasr}
\title{Post audio file to UWebASR speech recognition engine and retrieve transcription}
\usage{
post_uwebasr(lang_model = "CZ", path_file, format = "plaintext")
}
\arguments{
\item{lang_model}{Language of speech recognition model (CZ or SK)}

\item{path_file}{Path to file to upload and recognize speech}

\item{format}{Format of output}
}
\description{
Post audio file to UWebASR speech recognition engine and retrieve transcription
}
