% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{uwebasr_get}
\alias{uwebasr_get}
\title{Get audio transcription from UWebASR speech recognition engine}
\usage{
uwebasr_get(lang_model, audio_url, format = "plaintext")
}
\arguments{
\item{lang_model}{Language of speech recognition model (CZ or SK)}

\item{audio_url}{URL to audio file}

\item{format}{Format of output}
}
\description{
Get audio transcription from UWebASR speech recognition engine
}
